refactor: database structure and prisma integration
- Removed the generated Prisma schema and related files, they are now in node_modules and imported by `@prisma/client`
- Rename `prisma-client.ts` to `prisma-manager.ts` to avoid ambiguity and update related imports in queies
- Changed the output path for the Prisma client in the schema file to align with the new structure.
- Updated all database query files and IPC channel types to import directly from `@prisma/client`.
