+++ src/server/database/queries/formula-conversations-queies.ts
@@ -1,6 +1,6 @@
-import { PrismaClient } from '../generated';
-import { getPrismaClient, saveData } from '../prisma-client';
-import type { Prisma, formula_conversations } from '../generated';
+import { PrismaClient } from '@prisma/client';
+import { getPrismaClient, saveData } from '../prisma-manager';
+import type { Prisma, formula_conversations } from '@prisma/client';
 
 /**
  * `formula_conversations` table queries, only contains static methods
