+++ src/server/database/queries/formula-tags-queies.ts
@@ -1,6 +1,6 @@
-import { PrismaClient } from '../generated';
-import { getPrismaClient, saveData } from '../prisma-client';
-import type { formula_tags } from '../generated';
+import { PrismaClient } from '@prisma/client';
+import { getPrismaClient, saveData } from '../prisma-manager';
+import type { formula_tags } from '@prisma/client';
 
 /**
  * `formula_tags` table queries, only contains static methods
